Drop the future library
We no longer need it having dropped support for Python 2.
