+++ setup.cfg
@@ -2,7 +2,7 @@
 test = pytest
 
 [flake8]
-exclude = .*,env,lib,dist,build,tests/django_app,tests/py3*
+exclude = .*,env,lib,dist,build,tests/django_app
 ignore = E402,W503
 
 [tool:pytest]
@@ -48,7 +48,6 @@ install_requires =
 	statsd<4;python_version>="3.6"
 	requests~=2.25;python_version~="3.5.0"
 	requests<3.0;python_version>"3.5"
-	future~=0.18
 	contextvars~=2.4;python_version>="3.5" and python_version<"3.7"
 
 [options.extras_require]
