+++ talisker/requests.py
@@ -30,14 +30,13 @@ import logging
 import random
 import warnings
 import time
-from future.moves.urllib.parse import (
+from urllib.parse import (
     parse_qsl,
     urlparse,
     urlsplit,
     urlunsplit,
 )
 
-import future.utils
 import requests
 from requests.adapters import HTTPAdapter
 import requests.exceptions
@@ -66,7 +65,7 @@ __all__ = [
 STORAGE = Local()
 STORAGE.sessions = {}
 HOSTS = module_dict()
-DEBUG_HEADER = future.utils.text_to_native_str('X-Debug')
+DEBUG_HEADER = 'X-Debug'
 
 
 def clear():
@@ -501,19 +500,7 @@ class TaliskerAdapter(HTTPAdapter):
                     raise
 
             if retries_exhausted:
-                # An interesting bit of py2/3 differences here. We want to
-                # reraise the original ConnectionError, with context unaltered.
-                # A naked raise does exactly this in py3, it reraises the
-                # exception that caused the current except: block.  But in py2,
-                # naked raise would raise the *last* error, MaxRetryError,
-                # which we do not want. So we explicitly reraise the original
-                # ConnectionError. In py3, this would not be desirable, as the
-                # exception context would be confused, by py2 does not do
-                # chained exceptions, so, erm, yay?
-                if future.utils.PY3:
-                    raise  # raises the original ConnectionError
-                else:
-                    raise exc
+                raise  # raises the original ConnectionError
 
             # we are going to retry, so backoff as appropriate
             request._retry.sleep()
